% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{bpi}
\alias{bpi}
\title{bathytopografic position index}
\usage{
bpi(dem, orad, irad = NULL, slope = NULL, fac = 1, simple = TRUE,
  simpler = TRUE, scale = FALSE, inv = FALSE, verbose = FALSE,
  raw = FALSE, ...)
}
\arguments{
\item{dem}{the elevation model as \code{RasterLayer}, \code{SpatialGridDataFrame} or \code{SpatialPointsDataFrame}}

\item{orad}{outer circle radius in map units for which the tpi is calculated}

\item{irad}{inner circle radius. If given, an annulus is used fot tpi caculation. If 0, the inner radius contains exacly one pixel. Default is \code{NULL}}

\item{slope}{threshold to delineate flat from slope. Unit is 'degrees'. In case of default \code{NULL}, single standart deviation of slopes is used}

\item{fac}{modifier to slope threshold. Applies only if \code{simple=TRUE}.}

\item{simple}{output only 3 categories: (1) valley; (2) slope/flat; (3) ridge/top. Overrides \code{simpler}}

\item{simpler}{simplify the non-simple schema to 6 categories:  (1) valley; (2) lower slope; (3) flat; (4) middle slope; (5) upper slope; (6) ridge/top}

\item{scale}{apply preemptive \code{\link{scale}}}

\item{inv}{switch sign (in case of bathemetry with positive depth values)}

\item{verbose}{be verbose and plot}
}
\description{
Returns different tpi's.
}
\details{
#FIXME
}
\examples{
#FIXME

}

