% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{crecs}
\alias{crecs}
\alias{crecs-package}
\title{crecs: Comfy Multivariate Regression and Cross-validation with Spatial Objects}
\description{
crecs provides wrapper functions for some popular multivariate regression and catgorization / species distribution modelling methods.
}
\details{
It has three main goals:
\itemize{
\item Free you from tedious tasks trough easy automation.
\item Provide a large variety of model performance metrics.
\item Provide model validation techniques like cross-validation, null-models,
and  exploration of spatial autocorrelation.
}
}

