% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crossvalid}
\alias{crossvalid}
\title{Cross validation routine for species distribution models}
\usage{
crossvalid(Ncv, Kfold, data, method, responsetype, response, predictors,
  secondary = NULL, strata = NULL, enviStack, enviPix, seed,
  check.names = TRUE, aggregated = NULL, pseudoabsence = NULL,
  gbm.trees = 2000, maxargs = c("outputformat=logistic",
  "defaultprevalence=0.5"), rast = TRUE, flat = FALSE, ...)
}
\arguments{
\item{Ncv}{Number of cv repititions.}

\item{Kfold}{Number of folds. Usually take 10 or 5.}

\item{data}{\code{SpatialPointsDataFrame} containing response and predictors}

\item{method}{SDM methos used: "gam", "rf", "gbm", "max", or "gbm.step". See details for details}

\item{responsetype}{Type of the response: "count", "continuous", or "presence". Presence denotes bimodal responses [0|1]}

\item{response}{Column name of response in data argument}

\item{predictors}{Column names of predictors to use in data argument}

\item{secondary}{Column name in data. Use this to calculate model performance metrics from instead of response. Default is NULL.}

\item{strata}{criterion for fold generation. #FIXME}

\item{enviStack}{\code{RasterStack} of predictors. Used to calculate SD map}

\item{enviPix}{\code{SpatialPixelsDataFrame} of predictors. enviPix<-as(enviStack,"SpatialPixelsDataFrame"). Only for performance.}

\item{seed}{Integer. You probably want reproduceability. Note that Maxent's pseudoabsence generation can't be seeded this way--so expect those results to vary}

\item{aggregated}{logical or \code{NULL} Default is  \code{NULL}. This is for ensemble calculations and added to metrics in case of not being \code{NULL}}

\item{pseudoabsence}{logical or \code{NULL} Default is \code{NULL}. Same as above}

\item{gbm.trees}{gbm.trees param for dismo::gbm}

\item{maxargs}{argument to pass tp maxent}

\item{rast}{return \code{list} of metric and raster}

\item{flat}{return model performance metrics only (as \code{data.frame})}

\item{...}{ellipsis is used to pass arguments to subsequent functions like  \code{threshold.def}. See \code{\link{metrics}} for details}
}
\description{
repeated k-fold cross-validation. Calculate model preformance metrics, disribution, standard deviation and occurence maps.
}

