% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{deutschebucht}
\alias{deutschebucht}
\title{Evironmental conditions in the Deutsche Bucht}
\format{a \code{RasterBrick} containing 8 layers.}
\source{
various
}
\usage{
data(deutschebucht)
}
\description{
his data set gives the depth, median grain size, sorting coefficient,
mud content, average shear stress by currents, average shear stress by waves,
winter bottom water temperature, and average primary production of
the south-east sector of the North Sea, the Deutsche Bucht.
}
\author{
Jan M Holstein , 2015-03-05
}

