% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.dataset.R
\name{get.environ}
\alias{get.environ}
\title{read predictors and add respective values to your spatial points}
\usage{
get.environ(points, environ, dropcols = TRUE)
}
\arguments{
\item{points}{\code{SpatialPoints} or \code{SpatialPointsDataFrame}}

\item{environ}{\code{RasterLayer}, \code{RasterStack}, or \code{RasterBrick}}

\item{dropcols.}{if \code{TRUE}, remove preexisting columns in \code{points} that also exist in \code{environ} under the same name.}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
takes your response as \code{SpatialPoints} or \code{SpatialPointsDataFrame} and add  values from  
a \code{RasterLayer}, \code{RasterStack}, or \code{RasterBrick} at those positions. New columns will have the names of the respective raster layers.
}
\details{
If \code{FALSE}, preexisting columns in \code{points} will be overwritten by column creates by looking up values in raster layers with the same name.
#FIXME
}
\examples{
#FIXME

}

