% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lfi}
\alias{lfi}
\title{landform analysis indices}
\usage{
lfi(dem, large.bpi, small.bpi, slope.threshold = NULL, factors = TRUE,
  verbose = FALSE, trim = F)
}
\arguments{
\item{dem}{the elevation model as \code{RasterLayer}}

\item{large.bpi}{large scale \code{\link{bpi}} raster. Must use \code{simple=TRUE}}

\item{small.bpi}{small scale \code{\link{bpi}} raster. Must use \code{simple=TRUE}}

\item{factors}{make factors (also include landform names)}

\item{verbose}{be verbose and plot}

\item{trim}{trim the output raster to non-NA extent}

\item{slope}{threshold to delineate flat from slope. Unit is 'degrees'. In case of default \code{NULL}, single standart deviation of slopes is used}
}
\description{
Returns landform analyses
}
\details{
#FIXME
#FIXME
}
\examples{
#FIXME
}

