% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model}
\alias{model}
\title{Make a species distribution model}
\usage{
model(data, method, responsetype, response, predictors, secondary = NULL,
  enviStack, enviPix, seed = NULL, aggregated = NULL,
  pseudoabsence = NULL, gbm.trees = 2000,
  maxargs = c("outputformat=logistic", "defaultprevalence=0.5"),
  model = FALSE, prediction = FALSE, flat = FALSE, moran = FALSE,
  rast = TRUE, rast.occ = TRUE, ...)
}
\arguments{
\item{data}{\code{SpatialPointsDataFrame} containing response and predictors}

\item{method}{SDM methos used: "gam", "rf", "gbm", "max", or "gbm.step". See details for details}

\item{responsetype}{Type of the response: "count", "continuous", or "presence". Presence denotes bimodal responses [0|1]}

\item{response}{Column name of response in data argument}

\item{predictors}{Column names of predictors to use in data argument}

\item{secondary}{Column name in data. Use this to calculate model performance metrics from instead of response. Default is NULL.}

\item{enviStack}{\code{RasterStack} of predictors. Used to calculate SD map}

\item{enviPix}{\code{SpatialPixelsDataFrame} of predictors. enviPix<-as(enviStack,"SpatialPixelsDataFrame"). Only for performance.}

\item{seed}{Integer. For reproduceability.}

\item{aggregated}{logical or \code{NULL} Default is  \code{NULL}. This is for ensemble calculations and added to metrics in case of not being \code{NULL}}

\item{pseudoabsence}{logical or \code{NULL} Default is \code{NULL}. Same as above}

\item{gbm.trees}{gbm.trees param for dismo::gbm}

\item{maxargs}{argument to pass tp maxent}

\item{model}{\code{logical} return ONLY the model. Default is \code{FALSE}. Overrides args below.}

\item{prediction}{return ONLY the prediction. Default is \code{FALSE}. Overrides args below.}

\item{flat}{return model performance metrics only (as \code{data.frame})}

\item{moran}{add spatial autocorrelation metric to model metric. Default is \code{FALSE}. See details.}

\item{rast}{return \code{list} of metric and raster}

\item{rast.occ}{additionally adds bimodal occurence map. see details for threshold calculation.}

\item{...}{ellipsis is used to pass arguments to subsequent functions like  \code{threshold.def} and \code{moRange}. See \code{\link{metrics}} \code{\link{moranii}} for details}
}
\value{
A \code{model}, or a \code{data.frame} or a \code{list} depending on arguments
}
\description{
Returns the model, the performance metrics, and/or distribution maps depending on arguments.
}
\details{
SDM methos used are "gam", "rf", "gbm", "max", or "gbm.step".
If you want spatial autocorrelation metrics you probably need to pass additional arguments to \code{\link{moranii}}. Note that calculations may take very long depending on number of points and parametrization #FIXME
}
\examples{
#FIXME
}

