% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.dataset.R
\name{pseudo.absence}
\alias{pseudo.absence}
\title{sample background aka pseudo-absence data}
\usage{
pseudo.absence(presence, raster, area = NULL, sample.size = 1000,
  extent = "raster", buffer = NULL, chunks = NULL, verbose = TRUE,
  seed = NULL, xtracols = NULL, out = "spdf", colname = NULL)
}
\arguments{
\item{presence}{\code{SpatialPoints} of presence data}

\item{raster}{\code{RasterLayer} is required for sampling. If \code{buffer=NULL}, the background is sampled without the pixels with presence data.}

\item{area}{\code{SpatialPolygons} or  \code{SpatialPolygonsDataFrame} with an area within \code{raster} to be sampled (cmp. \code{extent})}

\item{sample.size}{The sample size}

\item{extent}{Eiter \code{raster(default), extent, or area}. Sample the whole area (extent), only where raster vals are !=NA (rater) or confined to an area (area).}

\item{buffer}{Perimeter around presence data where background is not sampled. Distance unit of buffer: see \code{\link{raster::buffer}}.}

\item{verbose}{Be verbose}

\item{seed}{Seed for random background sampling}

\item{xtracols}{transfer given features from \code{presence}--input data \code{SpatialPointsDataFrame} to returned \code{SpatialPointsDataFrame}.}

\item{out}{If \code{spdf}(default), a SpatialPointDataFrame is returned with a column "presence" with 1:presence 0:absence. If \code{sp}, only the \code{SpatialPoints} of the background are returned.  Overrides \code{xtracols}}

\item{colname}{New colnames for \code{xtracols}. \code{length(colname)==length(xtracols)}.}
}
\value{
\code{SpatialPointsDataFrame} or \code{SpatialPoints}. See \code{out}.
}
\description{
takes \code{SpatialPoints} or \code{SpatialPointsDataFrame} and add  values from  
FIXME
}
\details{
#FIXME
}
\examples{
#FIXME

}

