% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{species}
\alias{species}
\title{Virtual sampling of three virtual species from the Deutsche Bucht}
\format{a \code{SpatialPointsDataFrame} containing 3x300 observations.}
\source{
made up with virtualspecies package
}
\usage{
data(species)
}
\description{
his data set gives 300 random sampling locations within the Deutsche Bucht with
3 virtual species including sampling error.
}
\author{
Jan M Holstein , 2015-03-05
}
\seealso{
\code{\link{deutschebucht}}
}

