% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crecs.R
\docType{package}
\name{crecs}
\alias{crecs}
\alias{crecs-package}
\title{crecs: Comfy Multivariate Regression and Cross-validation with Spatial Objects}
\description{
crecs provides wrapper functions for some popular multivariate regression and catgorization / species distribution modelling methods.
It has three main goals:
\describe{
\itemize{
\item Free you from tedious tasks trough easy automation.
\item Provide a large variety of model performance metrics.
\item Provide model validation techniques like cross-validation, null-models, 
}
and  exploration of spatial autocorrelation.
}
}
\keyword{package}
\keyword{sdm}

